Implements full Lie normalization
Adds functionality for full Lie normalization, which eliminates all non-resonant terms from the Hamiltonian.

This contrasts with the partial normal form used in center manifold reduction, resulting in a sparser normal form containing only resonant terms.
